+++ harpy/_conda.py
@@ -8,7 +8,7 @@ from rich.table import Table
 from rich import print as rprint
 from ._printing import print_error, print_solution_with_culprits
 
-def create_conda_recipes(outdir: str, envs: list=None) -> None:
+def create_conda_recipes(outdir: str, envs: list= []) -> None:
     """Create the YAML files of the workflow conda dependencies"""
     environ = {
         "align" : [